datetime and time functions proviced in central util modul, provide name and description for dashobard widget type report
