+++ frontend/src/pages/GameDetails.tsx
@@ -48,6 +48,7 @@ import CameraAltIcon from '@mui/icons-material/CameraAlt';
 import YouTubeIcon from '@mui/icons-material/YouTube';
 import WeatherModal from '../modals/WeatherModal';
 import { WeatherDisplay } from '../components/WeatherIcons';
+import { formatEventTime, formatDateTime } from '../utils/formatter'
 
 interface GameDetailsProps {
   gameId: number;
@@ -192,20 +193,6 @@ function GameDetailsInner({ gameId, onBack }: GameDetailsProps) {
     } catch (err) {
       setError(err instanceof Error ? err.message : 'Fehler beim Laden der Ereignisse');
     }
-  };
-
-  const formatDateTime = (dateString: string) => {
-    const date = new Date(dateString);
-    const pad = (n: number) => n.toString().padStart(2, '0');
-    return `${pad(date.getUTCDate())}.${pad(date.getUTCMonth() + 1)}.${date.getUTCFullYear()} ${pad(date.getUTCHours())}:${pad(date.getUTCMinutes())}`;
-  };
-
-  const formatEventTime = (eventDate: string, gameStartDate: string) => {
-    const eventTime = new Date(eventDate);
-    const gameStart = new Date(gameStartDate);
-    const diffMs = eventTime.getTime() - gameStart.getTime();
-    const diffMinutes = Math.floor(diffMs / (1000 * 60));
-    return `${diffMinutes}'`;
   };
 
   const handleDeleteEvent = async () => {
