+++ frontend/src/pages/GameEvents.tsx
@@ -25,6 +25,7 @@ import {
 } from '../services/games';
 import { Game, GameEvent } from '../types/games';
 import { GameEventModal } from '../modals/GameEventModal';
+import { formatEventTime, formatDateTime } from '../utils/formatter';
 
 interface GameEventsProps {
   gameId: number;
@@ -59,20 +60,6 @@ export default function GameEvents({ gameId }: GameEventsProps) {
     } finally {
       setLoading(false);
     }
-  };
-
-  const formatDateTime = (dateString: string) => {
-    const date = new Date(dateString);
-    const pad = (n: number) => n.toString().padStart(2, '0');
-    return `${pad(date.getUTCDate())}.${pad(date.getUTCMonth() + 1)}.${date.getUTCFullYear()} ${pad(date.getUTCHours())}:${pad(date.getUTCMinutes())}`;
-  };
-
-  const formatEventTime = (eventDate: string, gameStartDate: string) => {
-    const eventTime = new Date(eventDate);
-    const gameStart = new Date(gameStartDate);
-    const diffMs = eventTime.getTime() - gameStart.getTime();
-    const diffMinutes = Math.floor(diffMs / (1000 * 60));
-    return `${diffMinutes}'`;
   };
 
   const getCurrentGameTime = () => {
